Move moduleService and ErrStopProcess to modules.
