+++ services/service.go
@@ -2,12 +2,8 @@ package services
 
 import (
 	"context"
-	"errors"
 	"fmt"
 )
-
-// ErrStopProcess is the error returned by a service as a hint to stop the server entirely.
-var ErrStopProcess = errors.New("stop process")
 
 // State of the service. See Service interface for full state diagram.
 type State int
